tripy/examples/segment-anything-model-v2/requirements.txt
-f https://nvidia.github.io/TensorRT-Incubator/packages.html
nvtripy
torch>=2.3.1
torchvision>=0.18.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
matplotlib>=3.9.1
jupyter>=1.0.0
opencv-python-headless>=4.7.0
